+++ UnitTests/Support/EventImplementingInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * @package Epa
+ */
+namespace Epa\UnitTests\Support;
+
+use Epa\Event;
+
+/**
+ * @package Epa
+ */
+class EventImplementingInterface implements Event, InterfaceWithDocCommentEvent
+{}