+++ UnitTests/Support/InterfaceWithDocCommentEvent.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * @package Epa
+ */
+namespace Epa\UnitTests\Support;
+
+/**
+ * @package Epa
+ * @event
+ * @eventName Foo
+ */
+interface InterfaceWithDocCommentEvent
+{}