+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Epa Changelog
 =============
 
+0.2.2
+
+	* Added a plugin that looks at the doc comments of interfaces that are implemented
+	  by an event. `@event` and `@eventName` allow to add custom event names.
+
 0.2.1
 
 	* EventDispatcher throws event `NewEventEvent` when notified of new event.